auth_api/requirements.txt
backoff==1.11.1
click==8.1.2
email-validator==1.1.3
flasgger==0.9.5
Flask==2.1.1
Flask-Babel==2.0.0
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
gunicorn==20.1.0
marshmallow-sqlalchemy==0.28.0
orjson==3.6.8
phonenumbers==8.12.46
psycopg2-binary==2.9.3
redis==4.2.2
SQLAlchemy==1.4.35
user-agents==2.2.0